Convert some buffer parameters to std::chrono. (5)
This commit focuses on converting some buffer parameters, and
propagating those changes throughout the rest of the code without
changing too many other interfaces. This removes some of the temporary
conversions to/from int/std::chrono::millisecond.

+++ mythtv/libs/libmythtv/mythplayerui.cpp
@@ -1030,7 +1030,7 @@ void MythPlayerUI::SwitchToProgram()
         delete ic;
     }
 
-    m_playerCtx->m_buffer->OpenFile(pginfo->GetPlaybackURL(), static_cast<uint>(MythMediaBuffer::kLiveTVOpenTimeout));
+    m_playerCtx->m_buffer->OpenFile(pginfo->GetPlaybackURL(), MythMediaBuffer::kLiveTVOpenTimeout);
 
     if (!m_playerCtx->m_buffer->IsOpen())
     {
@@ -1150,7 +1150,7 @@ void MythPlayerUI::JumpToProgram()
         delete ic;
     }
 
-    m_playerCtx->m_buffer->OpenFile(pginfo->GetPlaybackURL(), static_cast<uint>(MythMediaBuffer::kLiveTVOpenTimeout));
+    m_playerCtx->m_buffer->OpenFile(pginfo->GetPlaybackURL(), MythMediaBuffer::kLiveTVOpenTimeout);
     if (!m_playerCtx->m_buffer->IsOpen())
     {
         LOG(VB_GENERAL, LOG_ERR, LOC + QString("JumpToProgram's OpenFile failed (input type: %1)")
